+++ src/__tests__/utils/survey-event-receiver.test.ts
@@ -111,6 +111,27 @@ describe('survey-event-receiver', () => {
             expect(activatedSurveys).toContain('first-survey')
         })
 
+        it('receiver removes survey from list after its shown', () => {
+            const surveyEventReceiver = new SurveyEventReceiver(instance)
+            surveyEventReceiver.register(surveysWithEvents)
+            surveyEventReceiver.onEvent('billing_changed')
+            const activatedSurveys = surveyEventReceiver.getSurveys()
+            expect(activatedSurveys).toContain('first-survey')
+
+            surveyEventReceiver.onEvent('survey shown', {
+                $set: undefined,
+                $set_once: undefined,
+                event: 'survey shown',
+                timestamp: undefined,
+                uuid: '',
+                properties: {
+                    $survey_id: 'first-survey',
+                },
+            })
+
+            expect(surveyEventReceiver.getSurveys()).toEqual([])
+        })
+
         it('receiver activates same survey on multiple event', () => {
             const surveyEventReceiver = new SurveyEventReceiver(instance)
             surveyEventReceiver.register(surveysWithEvents)