+++ src/entrypoints/web-vitals.ts
@@ -1,0 +1,15 @@
+import { onLCP, onINP, onCLS, onFCP } from 'web-vitals'
+import { assignableWindow } from '../utils/globals'
+
+// TODO export types here as well?
+
+const postHogWebVitalsCallbacks = {
+    onLCP,
+    onCLS,
+    onFCP,
+    onINP,
+}
+
+assignableWindow.postHogWebVitalsCallbacks = postHogWebVitalsCallbacks
+
+export default postHogWebVitalsCallbacks