+++ src/posthog-surveys-types.ts
@@ -25,6 +25,8 @@ export interface SurveyAppearance {
     borderColor?: string
     position?: 'left' | 'right' | 'center'
     placeholder?: string
+    shuffleQuestions?: boolean
+    surveyPopupDelaySeconds?: number
     // widget options
     widgetType?: 'button' | 'tab' | 'selector'
     widgetSelector?: string
@@ -33,7 +35,6 @@ export interface SurveyAppearance {
     // questionable: Not in frontend/src/types.ts -> SurveyAppearance, but used in site app
     maxWidth?: string
     zIndex?: string
-    shuffleQuestions?: boolean
 }
 
 export enum SurveyType {
@@ -53,11 +54,7 @@ interface SurveyQuestionBase {
     optional?: boolean
     buttonText?: string
     originalQuestionIndex: number
-    branching?:
-        | NextQuestionBranching
-        | ConfirmationMessageBranching
-        | ResponseBasedBranching
-        | SpecificQuestionBranching
+    branching?: NextQuestionBranching | EndBranching | ResponseBasedBranching | SpecificQuestionBranching
 }
 
 export interface BasicSurveyQuestion extends SurveyQuestionBase {
@@ -94,7 +91,7 @@ export enum SurveyQuestionType {
 
 export enum SurveyQuestionBranchingType {
     NextQuestion = 'next_question',
-    ConfirmationMessage = 'confirmation_message',
+    End = 'end',
     ResponseBased = 'response_based',
     SpecificQuestion = 'specific_question',
 }
@@ -103,8 +100,8 @@ interface NextQuestionBranching {
     type: SurveyQuestionBranchingType.NextQuestion
 }
 
-interface ConfirmationMessageBranching {
-    type: SurveyQuestionBranchingType.ConfirmationMessage
+interface EndBranching {
+    type: SurveyQuestionBranchingType.End
 }
 
 interface ResponseBasedBranching {
@@ -159,6 +156,7 @@ export interface Survey {
             values: ActionType[]
         } | null
         events: {
+            repeatedActivation?: boolean
             values: {
                 name: string
             }[]